Resolve CHANGES.rst conflict: keep upstream's new Unreleased entries and
append the issue 2087 note plus its link target.

+++ tests/test_json.py
@@ -7,7 +7,6 @@ from __future__ import annotations
 
 import json
 import os
-
 from datetime import datetime
 from typing import Any
 
@@ -15,8 +14,7 @@ import pytest
 
 import coverage
 from coverage import Coverage
-
-from tests.coveragetest import UsingModulesMixin, CoverageTest
+from tests.coveragetest import CoverageTest, UsingModulesMixin
 
 
 class JsonReportTest(UsingModulesMixin, CoverageTest):
@@ -52,15 +50,13 @@ class JsonReportTest(UsingModulesMixin, CoverageTest):
             "num_statements": statements,
             "percent_covered": percent_covered,
             "percent_covered_display": (
-                f"{percent_covered:.{precision}f}"
-                if precision > 0
-                else str(int(round(percent_covered)))
+                f"{percent_covered:.{precision}f}" if precision > 0 else str(round(percent_covered))
             ),
             "percent_statements_covered": percent_statements_covered,
             "percent_statements_covered_display": (
                 f"{percent_statements_covered:.{precision}f}"
                 if precision > 0
-                else str(int(round(percent_statements_covered)))
+                else str(round(percent_statements_covered))
             ),
         }
 
@@ -81,7 +77,7 @@ class JsonReportTest(UsingModulesMixin, CoverageTest):
                     "percent_branches_covered_display": (
                         f"{percent_branches_covered:.{precision}f}"
                         if precision > 0
-                        else str(int(round(percent_branches_covered)))
+                        else str(round(percent_branches_covered))
                     ),
                 }
             )
